v2/utils/config.pyの除去
