+++ api/v2/cruds/user.py
@@ -6,7 +6,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 import v2.models.user as user_model
 import v2.schemas.user as user_schema
-from v2.utils.config import get_openai_api_key
+from config import get_openai_api_key
 
 client = OpenAI(api_key=get_openai_api_key())
